# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediTrack Pro API — PULSO IoT Pill Dispensing System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{patientId}/devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List IoT dispensers for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: pairDevice
      tags: [devices]
      summary: Pair a new dispenser
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevicePairInput"
      responses:
        "201":
          description: Paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /patients/{patientId}/devices/{deviceId}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get a device with funnel status
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device with funnels
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceDetail"
  /patients/{patientId}/devices/{deviceId}/dispense:
    post:
      operationId: manualDispense
      tags: [devices]
      summary: Manual dispense (PIN-gated)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispenseInput"
      responses:
        "200":
          description: Dispensed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispenseResult"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{patientId}/devices/{deviceId}/logs:
    get:
      operationId: getDeviceLogs
      tags: [devices]
      summary: Get device activity logs
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceLog"

  # ── Adherence ──────────────────────────────────────────────
...
  /patients/{patientId}/adherence/{doseEventId}/confirm:
    post:
      operationId: confirmDose
      tags: [adherence]
      summary: Confirm a dose was taken
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
        - name: doseEventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoseConfirmInput"
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoseEvent"
...
  /patients/{patientId}/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get full dashboard summary for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── Emergency QR ───────────────────────────────────────────
  /patients/{patientId}/emergency-qr:
    get:
      operationId: getEmergencyQr
      tags: [qr]
      summary: Get emergency QR data for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emergency QR data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyQrData"
  /emergency/{token}:
    get:
      operationId: getEmergencyPublicProfile
      tags: [qr]
      summary: Zero-login emergency profile (QR scan target)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public emergency profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyProfile"
        "404":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Notifications ──────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Auth
...
    DevicePairInput:
      type: object
      required: [deviceCode, ssid, wifiPassword]
      properties:
        deviceCode:
          type: string
        ssid:
          type: string
        wifiPassword:
          type: string
        nickname:
          type: string
    DispenseInput:
      type: object
      required: [medicationId, pin]
      properties:
        medicationId:
          type: integer
        pin:
          type: string
        funnelNumber:
          type: integer
    DispenseResult:
      type: object
      required: [success, dispensedAt]
      properties:
        success:
          type: boolean
        dispensedAt:
          type: string
          format: date-time
        message:
          type: string
    Funnel:
      type: object
      required: [number, medicationId, stockCount, status]
      properties:
        number:
          type: integer
        medicationId:
          type: ["integer", "null"]
        medicationName:
          type: ["string", "null"]
        stockCount:
          type: integer
        status:
          type: string
          enum: [ok, low, empty, jam]
    Device:
      type: object
      required: [id, patientId, status, nickname]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        nickname:
          type: string
        deviceCode:
          type: string
        status:
          type: string
          enum: [online, offline, error]
        wifiStrength:
          type: ["integer", "null"]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DeviceDetail:
      type: object
      required: [id, patientId, status, nickname, funnels]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        nickname:
          type: string
        deviceCode:
          type: string
        status:
          type: string
          enum: [online, offline, error]
        wifiStrength:
          type: ["integer", "null"]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
        funnels:
          type: array
          items:
            $ref: "#/components/schemas/Funnel"
        createdAt:
          type: string
          format: date-time
    DeviceLog:
      type: object
      required: [id, deviceId, event, timestamp]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        event:
          type: string
          enum: [dispense, missed, jam, refill, online, offline, error]
        message:
          type: string
        funnelNumber:
          type: ["integer", "null"]
        timestamp:
          type: string
          format: date-time

    # Adherence
    DoseEvent:
      type: object
      required: [id, patientId, medicationId, scheduledAt, status]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: string
        scheduledAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, taken, missed, snoozed]
    DoseConfirmInput:
      type: object
      required: [confirmedAt]
      properties:
        confirmedAt:
          type: string
          format: date-time
        method:
          type: string
          enum: [tap, voice, auto]
          default: tap
...
    TrendFlag:
      type: object
      required: [id, type, message, severity]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [weekend_misses, evening_misses, low_stock, streak_risk, perfect_week]
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        detectedAt:
          type: string
          format: date-time
...
    UpcomingDose:
      type: object
      required: [doseEventId, medicationId, medicationName, scheduledAt, status]
      properties:
        doseEventId:
          type: integer
          description: The underlying DoseEvent id — pass this to POST /adherence/{doseEventId}/confirm to mark it taken.
        medicationId:
          type: integer
        medicationName:
          type: string
        dosage:
          type: string
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, taken, missed, overdue]
        minutesUntilDue:
          type: ["integer", "null"]
    DashboardSummary:
      type: object
      required: [patientId, patientName, upcomingDoses, adherenceRate, currentStreak, stockAlerts, deviceStatus]
      properties:
        patientId:
          type: integer
        patientName:
          type: string
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        nextDoseAt:
          type: ["string", "null"]
          format: date-time
        minutesUntilNextDose:
          type: ["integer", "null"]
        upcomingDoses:
          type: array
          items:
            $ref: "#/components/schemas/UpcomingDose"
        adherenceRate:
          type: number
        currentStreak:
          type: integer
        dosesToday:
          type: integer
        dosesTakenToday:
          type: integer
        stockAlerts:
          type: array
          items:
            type: object
            properties:
              medicationName:
                type: string
              daysRemaining:
                type: integer
              funnelNumber:
                type: integer
        deviceStatus:
          type: string
          enum: [online, offline, error, unpaired]
        trendFlags:
          type: array
          items:
            $ref: "#/components/schemas/TrendFlag"

    # Emergency QR
    EmergencyQrData:
      type: object
      required: [patientId, token, qrUrl]
      properties:
        patientId:
          type: integer
        token:
          type: string
        qrUrl:
          type: string
        generatedAt:
          type: string
          format: date-time
    EmergencyProfile:
      type: object
      required: [name, bloodType, conditions, allergies, medications]
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        bloodType:
          type: string
        conditions:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        medications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              dosage:
                type: string
              form:
                type: string
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]

    # Notifications
...