# packages/api-zod/src/generated/types/deviceDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MediTrack Pro API — PULSO IoT Pill Dispensing System
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceDetailStatus } from './deviceDetailStatus';
import type { Funnel } from './funnel';

export interface DeviceDetail {
  id: number;
  patientId: number;
  nickname: string;
  deviceCode?: string;
  status: DeviceDetailStatus;
  /** @nullable */
  wifiStrength?: number | null;
  /** @nullable */
  lastSeenAt?: Date | null;
  funnels: Funnel[];
  createdAt?: Date;
}