# packages/api-zod/src/generated/types/dispenseResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * MediTrack Pro API — PULSO IoT Pill Dispensing System
 * OpenAPI spec version: 0.1.0
 */

export interface DispenseResult {
  success: boolean;
  dispensedAt: Date;
  message?: string;
}